pyproject.toml
[project]
name = "chunkviz"
version = "0.1.0"
description = "Visualize and evaluate document chunking strategies for RAG pipelines."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.8.2",
  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "pdfplumber>=0.11.4",
  "pymupdf>=1.24.9",
  "tiktoken>=0.7.0",
  "sentence-transformers>=3.0.1",
  "scikit-learn>=1.5.1",
  "streamlit>=1.36.0",
  "nltk>=3.9",
  "Pillow>=10.4.0"
]

[project.scripts]
chunkviz = "chunkviz.cli.app:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

# Explicit packages list — no auto discovery, no 'find' section.
[tool.setuptools]
packages = ["chunkviz"]
